+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 - Pathway will only require `LiteLLM` package, if you use one of the wrappers for `LiteLLM`.
+- Retries are implemented in `pw.io.airbyte.read`.
+- State processing protocol is updated in `pw.io.airbyte.read`.
 
 ## [0.8.3] - 2024-03-13